feat: extract query commands to analytics-query.ts
Move cmdStatus, cmdTrend, cmdDiff, cmdBaseline, cmdHistory from
analytics.ts to analytics-query.ts for better separation of concerns.
analytics.ts retains cmdInsights, cmdTasks, cmdReport, cmdHealth.

Closes #137
